django_app/pyproject.toml
[tool.poetry]
name = "django_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
django-prometheus = "^2.2.0"
python-decouple = "^3.5"
gunicorn = "^20.1.0"
django-health-check = "^3.16.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
